new colony sorting added
A new method, testSortColony() has been added to Controller.java. To enable this, the class PathComparator has also been added to the myUtils package. The testSortColony( ) method has been commented out at this stage, it is incomplete. The next task is to use the sorted colony of ants to be able to find the worst, second worst etc. (and best, second best etc.) ant in the colony for pheromone update. This is to enable experiments into increasing the strength of antipheromone (in readiness for self-adaption).