add second and third worst in colony to MMAS
In the daemonActions( ) method of the Controller class, after the paths in the colony have been sorted by Fcomb (using PathComparator), the second worst and third worst in the colony are now recorded. In the PheromoneOperators update() method, the MMAS update method has been changed to now include updates for the second and third worst solution paths in the colony.